+++ CHANGELOG.md
@@ -60,6 +60,7 @@ Added partial support for proposal-based asset updates:
 - Added methods:
   - propose_commit_batch() stores batch arguments for later commit
   - delete_batch() deletes a batch, intended for use after propose_commit_batch if cancellation needed
+  - compute_evidence() computes a hash ("evidence") over the proposed batch arguments
 
 Added `api_version` endpoint. With upcoming changes we will introduce breaking changes to asset canister's batch upload process. New endpoint will help `ic-asset` with differentiation between API version, and allow it to support all versions of the asset canister. 
 
@@ -67,7 +68,8 @@ Added `api_version` endpoint. With upcoming changes we will introduce breaking c
 
 ### Frontend canister
 
-- Module hash: a6476d0619d523ca5934120437474623c89534ffab997793aa76be003fca2144
+- Module hash: 1d170f9732b9c829a1484e7d68d74cf4a5447aa9df8b14df3c373cfda9e29ce0
+- https://github.com/dfinity/sdk/pull/3023
 - https://github.com/dfinity/sdk/pull/3022
 - https://github.com/dfinity/sdk/pull/3021
 - https://github.com/dfinity/sdk/pull/3019